grpc-auth-project/profile-service/requirements.txt
grpcio
grpcio-tools
python-dotenv # Only needed if you plan to use .env here too